Syncer add blocks to store instead of msg passing
Give the Syncer a `Store` so that it can add the blocks to the DB
directly, message passing does have advantages over this but this is
simpler. To be able to be notified when a new block was added the Store
also implements subscribing to blocks that have been added now using a
broadcast channel.